+++ release_notes.md
@@ -6,6 +6,7 @@ Spider2 Changelog
 
 ### New Feature
 * Adding new DEFAULT_NOSYNC allocator (which only work post scheduling) that removes synchronisation points on forks, duplicates and extern interfaces. Default remains DEFAULT allocator.
+* Adding full support of SR-DAG less execution of both static and dynamic applications.
 
 ### Changes
 * Changed Semaphore.h to use POSIX semaphore on Linux.